Typ von sizeof ist unter x86-64 nicht unsigned sondern "long unsigned". size_t sollte eigentlich funktionieren
git-svn-id: svn://svn.code.sf.net/p/tntnet/code/trunk@10 8dd45c8f-be11-0410-86c3-e0da43b70fc1

+++ tntnet/framework/common/http.cpp
@@ -209,7 +209,7 @@ void httpMessage::parseBody(std::istream& in)
 
     body.clear();
     char buffer[512];
-    unsigned size = content_size;
+    size_t size = content_size;
     while (size > 0
       && (in.read(buffer, std::min(sizeof(buffer), size)), in.gcount() > 0))
     {
